+++ www/api/OGDocsComponent.tsx
@@ -1,0 +1,46 @@
+import React from 'react';
+
+export const OGDocsComponent = ({
+  title,
+  description,
+}: {
+  title: string;
+  description: string;
+}) => {
+  return (
+    <div
+      style={{
+        height: '100%',
+        display: 'flex',
+        width: '100%',
+        flexDirection: 'column',
+        alignItems: 'center',
+        justifyContent: 'center',
+        position: 'relative',
+        fontFamily: 'Inter',
+      }}
+    >
+      <img
+        src="https://assets.trpc.io/www/og-pattern-light.svg"
+        tw="absolute"
+      />
+      <div tw="flex items-center justify-between h-full w-full p-16">
+        <div tw="flex flex-col items-start">
+          <h1 tw="text-6xl text-gray-900">{title}</h1>
+          <p tw="text-2xl text-gray-700">{description}</p>
+        </div>
+
+        <div tw="flex justify-between">
+          <div tw="flex flex-col items-center">
+            <img
+              tw="h-64 w-64 mb-4"
+              src="https://assets.trpc.io/icons/svgs/blue-bg-rounded.svg"
+              alt=""
+            />
+            <span tw="text-5xl font-bold">tRPC</span>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
