+++ www/api/ogDocs.ts
@@ -1,6 +1,6 @@
 import { ImageResponse } from '@vercel/og';
 import { z } from 'zod';
-import { OGDocsComponent } from './ogComponents';
+import { OGDocsComponent } from './OGDocsComponent';
 
 export const config = {
   runtime: 'experimental-edge',
